Crate nautilus_system

System-level components and orchestration for PoseiTrader.

The system crate provides the core system architecture for orchestrating trading systems, including the kernel that manages all engines, configuration management, and system-level factories for creating components:

PoseiKernel - Core system orchestrator managing engines and components.

PoseiKernelConfig - Configuration for kernel initialization.

System builders and factories for component creation.

Platform

PoseiTrader is an open-source, high-performance, production-grade algorithmic trading platform, providing quantitative traders with the ability to backtest portfolios of automated trading strategies on historical data with an event-driven engine, and also deploy those same strategies live, with no code changes.

PoseiTrader’s design, architecture, and implementation philosophy prioritizes software correctness and safety at the highest level, with the aim of supporting mission-critical, trading system backtesting and live deployment workloads.

Feature flags

This crate provides feature flags to control source code inclusion during compilation, depending on the intended use case, i.e. whether to provide Python bindings for the posei_trader Python package, or as part of a Rust only build.

python: Enables Python bindings from PyO3.

Re-exports

pub use builder::PoseiKernelBuilder;

pub use config::PoseiKernelConfig;

pub use factories::ClientConfig;

pub use factories::DataClientFactory;

pub use factories::ExecutionClientFactory;

pub use kernel::PoseiKernel;

Modules

builder

config

factories

kernel

trader

Central orchestrator for managing actors, strategies, and execution algorithms.