nautilus_system::config

Trait PoseiKernelConfig

pub trait PoseiKernelConfig: Debug {

    // Required methods
    fn environment(&self) -> Environment;
    fn trader_id(&self) -> TraderId;
    fn load_state(&self) -> bool;
    fn save_state(&self) -> bool;
    fn logging(&self) -> LoggerConfig;
    fn instance_id(&self) -> Option<UUID4>;
    fn timeout_connection(&self) -> u32;
    fn timeout_reconciliation(&self) -> u32;
    fn timeout_portfolio(&self) -> u32;
    fn timeout_disconnection(&self) -> u32;
    fn timeout_post_stop(&self) -> u32;
    fn timeout_shutdown(&self) -> u32;
    fn cache(&self) -> Option<CacheConfig>;
    fn msgbus(&self) -> Option<MessageBusConfig>;
    fn data_engine(&self) -> Option<DataEngineConfig>;
    fn risk_engine(&self) -> Option<RiskEngineConfig>;
    fn exec_engine(&self) -> Option<ExecutionEngineConfig>;
    fn portfolio(&self) -> Option<PortfolioConfig>;
    fn streaming(&self) -> Option<StreamingConfig>;

}

Configuration trait for a PoseiKernel core system instance.

Required Methods

fn environment(&self) -> Environment

Returns the kernel environment context.

fn trader_id(&self) -> TraderId

Returns the trader ID for the node.

fn load_state(&self) -> bool

Returns if trading strategy state should be loaded from the database on start.

fn save_state(&self) -> bool

Returns if trading strategy state should be saved to the database on stop.

fn logging(&self) -> LoggerConfig

Returns the logging configuration for the kernel.

fn instance_id(&self) -> Option<UUID4>

Returns the unique instance identifier for the kernel.

fn timeout_connection(&self) -> u32

Returns the timeout (seconds) for all clients to connect and initialize.

fn timeout_reconciliation(&self) -> u32

Returns the timeout (seconds) for execution state to reconcile.

fn timeout_portfolio(&self) -> u32

Returns the timeout (seconds) for portfolio to initialize margins and unrealized pnls.

fn timeout_disconnection(&self) -> u32

Returns the timeout (seconds) for all engine clients to disconnect.

fn timeout_post_stop(&self) -> u32

Returns the timeout (seconds) after stopping the node to await residual events before final shutdown.

fn timeout_shutdown(&self) -> u32

Returns the timeout (seconds) to await pending tasks cancellation during shutdown.

fn cache(&self) -> Option<CacheConfig>

Returns the cache configuration.

fn msgbus(&self) -> Option<MessageBusConfig>

Returns the message bus configuration.

fn data_engine(&self) -> Option<DataEngineConfig>

Returns the data engine configuration.

fn risk_engine(&self) -> Option<RiskEngineConfig>

Returns the risk engine configuration.

fn exec_engine(&self) -> Option<ExecutionEngineConfig>

Returns the execution engine configuration.

fn portfolio(&self) -> Option<PortfolioConfig>

Returns the portfolio configuration.

fn streaming(&self) -> Option<StreamingConfig>

Returns the configuration for streaming to feather files.